Replace sf.net with sourceforge.net
When I clicked on the link to signup for the mailing list Google Chrome
told me "Your connection is not private". That's because their SSL cert
is signed for *.sourceforge.net rather than *.sf.net. Yes, SourceForge
should have a separate cert for the sf.net domain but they don't and I
doubt they're going to fix it anytime soon.