Entirely remove Bootstrap and JQuery
This replaces Bootstrap with our own "solution".

We only used it for two things:

- Some very minor CSS that we could also change
- The "Download Fish" tab menu

The latter used javascript to select the tab, and fell back to
displaying *all platforms* if js was disabled.

We replace it with a CSS solution based on hidden radiobuttons
triggered by labels and some :checked magic. The entire menu now runs
without javascript, it's only necessary at start to select the current
platform - now it will fall back to selecting "Sources", and the
user'll have to select their platform (but they can!)

As to the CSS I've copied the most important parts, there are some
miniscule rendering differences - a pixel here and there, but nothing
noticeable unless you're switching back and forth side-by-side.

The button hover effect has been replaced with a simpler one, so
they'll just be a flat grey instead of a barely visible gradient.