Fix old pixelmator files
The older handful of .pxm documents are not possible to open in
in Pixelmator Pro versions from the last few years: "The document is
damaged, or isn’t compatible with Pixelmator Pro 2.2"

Apparently at some point compatiblity with Pixelmator's format circa
2012 was dropped.

The solution was to grab a more recent version of classic Pixelmator,
open the files and re-save them. Now everything can open them.