Make flex-basis smaller
This would make the image too wide, which would make the actual
suggestion part of the autosuggestion image invisible.

So we make the size that flex will try to give it smaller, so it gets
more height relative to the width (aspect ratios and such).

What we *should* do is replace the autosuggestion image with one that
isn't as high - basically try to find images with similar aspect ratios.