Add useful alt text
When it's just thumbnails having "thing thumbnail" is fine (tho the
images themselves never had any alt text, but given that we now show
them and they contain information, we should present that in the alt
text as well.

Or we should get the information out into the actual text - the
autosuggestions image shows a bunch of explanation.