Remove explicit linebreaks
This is a bit awkward when viewing the site at e.g. 1920px wide on a
1080p screen:

```
Finally, a command ................... LOGO
line shell for the 90s ............... LOGO
```

where the rest of the text would fit without a problem.

This will properly break and won't grow unbounded, so we can just let
it use two lines for the tagline and the sub-tagline (is that the term?)