Remove Comic Sans
I think this font is quite hard to read and makes the project look
like a joke; fish is fun but not a joke. The tongue-in-cheek mood
works well for people who are already fans of fish but I think the
website should focus on how it is perceived by outsiders.

(Related discussion in #127)

Closes #129